exposed tmpl, styles
